fix(stl): fix for using StlPySolver with JAX (temporary conversion to numpy)
